+++ developer/src/kmcmplib/commands.inc.sh
@@ -1,4 +1,7 @@
-#!/usr/bin/env bash
+# shellcheck shell=bash
+# no hashbang for .inc.sh
+
+. "$KEYMAN_ROOT/resources/locate_emscripten.inc.sh"
 
 # ----------------------------------------------------------------------------
 # clean
@@ -114,34 +117,6 @@ do_command() {
 # ----------------------------------------------------------------------------
 # utility functions
 # ----------------------------------------------------------------------------
-
-#
-# We don't want to rely on emcc being on the path, because Emscripten puts far
-# too many things onto the path (in particular for us, node).
-#
-# The following comment suggests that we don't need emcc on the path.
-# https://github.com/emscripten-core/emscripten/issues/4848#issuecomment-1097357775
-#
-# So we try and locate emcc in common locations ourselves. The search pattern
-# is:
-#
-# 1. Look for $EMSCRIPTEN_BASE (our primary emscripten variable), which should
-#    point to the folder that emcc is located in
-# 2. Look for $EMCC which should point to the emcc executable
-# 3. Look for emcc on the path
-#
-locate_emscripten() {
-  if [[ -z ${EMSCRIPTEN_BASE+x} ]]; then
-    if [[ -z ${EMCC+x} ]]; then
-      local EMCC=`which emcc`
-      [[ -z $EMCC ]] && builder_die "locate_emscripten: Could not locate emscripten (emcc) on the path or with \$EMCC or \$EMSCRIPTEN_BASE"
-    fi
-    [[ -x $EMCC ]] || builder_die "locate_emscripten: Variable EMCC ($EMCC) does not point to a valid executable emcc"
-    EMSCRIPTEN_BASE="$(dirname "$EMCC")"
-  fi
-
-  [[ -x ${EMSCRIPTEN_BASE}/emcc ]] || builder_die "locate_emscripten: Variable EMSCRIPTEN_BASE ($EMSCRIPTEN_BASE) does not point to emcc's folder"
-}
 
 build_meson_cross_file_for_wasm() {
   if [ $BUILDER_OS == win ]; then