+++ linux/scripts/package-build.inc.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 
 function checkPrerequisites() {
-    if [ "${UPLOAD}" == "yes" ]; then
+    if [ "${UPLOAD:=}" == "yes" ]; then
         SIM=""
     else
         SIM="-s"
     fi
 
     # Check the tier
-    if [[ -z "${TIER}" ]]; then
+    if [[ -z "${TIER:=}" ]]; then
         echo "TIER.md or \${TIER} must be set to (alpha, beta, stable) to use this script"
         exit 1
     fi
@@ -18,23 +18,17 @@ function checkPrerequisites() {
         exit 1
     fi
 
-    if [ "${PROJECT}" != "" ]; then
-        projects="${PROJECT}"
-    else
-        projects="keyman"
-    fi
+    # shellcheck disable=SC2034
+    projects="${PROJECT:=keyman}"
 }
 
 function downloadSource() {
     local packageDir
     packageDir=$1
 
-    if [ "${proj}" == "keyman" ]; then
+    if [ "${proj:=}" == "keyman" ]; then
        cd "${BASEDIR}" || exit
-    fi
-
-    if [ "${proj}" == "keyman" ]; then
-        make clean
+        ./build.sh clean
     fi
 
     # Update tier in Debian watch files (replacing any previously set tier) and remove comment
@@ -51,7 +45,7 @@ function downloadSource() {
     mv "${proj}"*.asc "${BASEDIR}/${packageDir}"
     rm "${proj}"*.debian.tar.xz
     cd "${BASEDIR}/${packageDir}" || exit
-    wget -N https://downloads.keyman.com/linux/${TIER}/${dirversion}/SHA256SUMS
+    wget -N "https://downloads.keyman.com/linux/${TIER}/${dirversion}/SHA256SUMS"
     sha256sum -c --ignore-missing SHA256SUMS |grep "${proj}"
 }